use chi instead of plain net/http
